% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atorusprez-package.R
\docType{package}
\name{atorusprez-package}
\alias{atorusprez}
\alias{atorusprez-package}
\title{atorusprez: Presentations with Atorus Style}
\description{
Generate presentations with Atorus Research styling, beginning with 'PowerPoint' but eventually expanding to 'reveal.js'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/atorusprez}
  \item \url{https://jonthegeek.github.io/atorusprez/}
  \item Report bugs at \url{https://github.com/jonthegeek/atorusprez/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jon.harmon@atorusresearch.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Other contributors:
\itemize{
  \item Atorus Research [copyright holder, funder]
}

}
\keyword{internal}
