% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pptx.R
\name{render_pptx}
\alias{render_pptx}
\title{Render qmd as pptx}
\usage{
render_pptx(path, date = Sys.Date(), quiet = TRUE, ...)
}
\arguments{
\item{path}{The path to the \code{qmd} file to render. By default, a new file will
be created in the same directory as \code{path}, with extension \code{.pptx}.}

\item{date}{A date to include on the title slide of the deck. Set this to
\code{NULL} if you do not wish to include a date.}

\item{quiet}{Whether to suppress output during the rendering process. Passed
on to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{...}{
  Arguments passed on to \code{\link[quarto:quarto_render]{quarto::quarto_render}}
  \describe{
    \item{\code{output_file}}{The name of the output file. If using \code{NULL}, the output
filename will be based on the filename for the input file. \code{output_file} is
mapped to the \code{--output} option flag of the \code{quarto} CLI. It is expected to
be a filename only, not a path, relative or absolute.}
    \item{\code{execute}}{Whether to execute embedded code chunks.}
    \item{\code{execute_params}}{A list of named parameters that override custom params
specified within the YAML front-matter.}
    \item{\code{execute_daemon}}{Keep Jupyter kernel alive (defaults to 300 seconds).
Note this option is only applicable for rendering Jupyter notebooks or
Jupyter markdown.}
    \item{\code{execute_daemon_restart}}{Restart keepalive Jupyter kernel before render.
Note this option is only applicable for rendering Jupyter notebooks or
Jupyter markdown.}
    \item{\code{execute_debug}}{Show debug output for Jupyter kernel.}
    \item{\code{use_freezer}}{Force use of frozen computations for an incremental
file render.}
    \item{\code{cache_refresh}}{Force refresh of execution cache.}
    \item{\code{metadata}}{An optional named list used to override YAML
metadata. It will be passed as a YAML file to \code{--metadata-file} CLI flag.
This will be merged over \code{metadata-file} options if both are
specified.}
    \item{\code{metadata_file}}{A yaml file passed to \code{--metadata-file} CLI flags to
override metadata. This will be merged with \code{metadata} if both are
specified, with low precedence on \code{metadata} options.}
    \item{\code{debug}}{Leave intermediate files in place after render.}
    \item{\code{profile}}{\href{https://quarto.org/docs/projects/profiles.html}{Quarto project profile(s)} to use. Either
a character vector of profile names or \code{NULL} to use the default profile.}
    \item{\code{quarto_args}}{Character vector of other \code{quarto} CLI arguments to append
to the Quarto command executed by this function. This is mainly intended for
advanced usage and useful for CLI arguments which are not yet mirrored in a
dedicated parameter of this \R function. See \verb{quarto render --help} for options.}
    \item{\code{as_job}}{Render as an RStudio background job. Default is \code{"auto"},
which will render individual documents normally and projects as
background jobs. Use the \code{quarto.render_as_job} \R option to control
the default globally.}
  }}
}
\value{
\code{NULL} invisibly (this function is called to generate the pptx as a
side effect).
}
\description{
Render a qmd as pptx, applying the Atorus theme.
}
